Moved shared translate function do separate file to avoid circular dependency
