+++ widgets/executions/src/SystemWorkflowIcon.tsx
@@ -1,5 +1,5 @@
 import type { Execution } from 'app/utils/shared/ExecutionUtils';
-import { translate } from './widget';
+import { translate } from './widget.utils';
 
 export default function SystemWorkflowIcon({ execution }: { execution?: Pick<Execution, 'is_system_workflow'> }) {
     const { Icon, Popup } = Stage.Basic;
