+++ widgets/executions/src/widget.utils.ts
@@ -1,0 +1,1 @@
+export const translate = Stage.Utils.getT('widgets.executions');
